pyproject.toml
[project]
name = "kings-paradox"
version = "0.1.0"
description = "The Crown's Paradox - A neuro-symbolic LLM-powered political intrigue game"
authors = [
    {name = "Kavin Stewart"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "dspy (>=3.0.4,<4.0.0)",
    "openai (>=2.14.0,<3.0.0)",
    "anthropic (>=0.75.0,<0.76.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "rich (>=14.2.0,<15.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "kings_paradox", from = "src"}]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
